requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
packaging==25.0
pdfminer-six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
pypdfium2==5.2.0
pytesseract==0.3.13
python-dotenv==1.2.1
